requirements.txt
fitz==0.0.1.dev2
Flask==2.2.5
nest_asyncio==1.6.0
numpy==1.21.6
numpy==1.24.3
pandas==1.3.5
pdfplumber==0.5.27
pdfplumber==0.11.0
tabula_py==2.6.0
tabula_py==2.9.3
Werkzeug==2.2.3
